% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{combineTCR}
\alias{combineTCR}
\title{combineTCR}
\usage{
combineTCR(
  df,
  df_clono,
  samples = NULL,
  ID = NULL,
  cells = "T-AB",
  removeNA = FALSE,
  removeMulti = FALSE,
  filterMulti = FALSE
)
}
\arguments{
\item{df}{List of filtered contig annotations from 10x Genomics.}

\item{df_clono}{List of clonotypes metadata from 10x Genomics.}

\item{samples}{The labels of samples (required).}

\item{ID}{The additional sample labeling (optional).}

\item{cells}{The type of T cell - T cell-AB or T cell-GD. Only 1 T cell type can
be called at once.}

\item{removeNA}{This will remove any chain without values.}

\item{removeMulti}{This will remove barcodes with greater than 2 chains.}

\item{filterMulti}{This option will allow for the selection of the 2
corresponding chains with the highest expression for a single barcode.}
}
\value{
List of clonotypes for individual cell barcodes
}
\description{
combineTCR
}
