% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{getTCR}
\alias{getTCR}
\title{getTCR}
\usage{
getTCR(vdj_dir, filteredcontigs = TRUE, removeNA = FALSE, removeMulti = FALSE)
}
\arguments{
\item{vdj_dir}{Path to 10X Genomics Cellranger vdj (or multi) output of VDJ contigs}

\item{removeNA}{Remove any chain with without values}

\item{removeMulti}{Remove barcodes with more than 2 chains}

\item{contigfile}{Contig info (10X Genomics VDJ output: filtered_contig_annotations.csv or all_filtered_contig_annotations.csv)}

\item{clonotypefile}{Clonotype metadata info(10X Genomics VDJ output: clonotypes.csv)}

\item{filterMulti}{This option will allow for the selection of the 2
corresponding chains with the highest expression for a single barcode.}
}
\value{
Data frame of clonotypes for individual cell barcodes

Example dataset: https://www.10xgenomics.com/resources/datasets/human-t-cells-from-a-healthy-donor-1-k-cells-multi-v-2-2-standard-5-0-0
https://www.10xgenomics.com/resources/datasets/human-pbmc-from-a-healthy-donor-10-k-cells-multi-v-2-2-standard-5-0-0
}
\description{
getTCR
}
\details{
Reads output directory of Cellranger VDJ (or multi) to output clonotype metadata (TCR chains, VDJ genes, reads, umis).
If extra info on T cell metadata is available (clonotypes.csv), it will be joined to the clonotype metadata.
}
\examples{
getTCR(vdj_dir="inst/extdata/1kcells-multi/vdj_t/")
getTCR(vdj_dir="inst/extdata/1kcells-multi/vdj_t/", filteredcontigs = FALSE, removeMulti = TRUE)
getTCR(contigfile="inst/extdata/1kcells-multi/vdj_t/filtered_contig_annotations.csv", clonotypefile="inst/extdata/1kcells-multi/vdj_t/clonotypes.csv")
}
