% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding.R
\name{plot_embed_clonotype}
\alias{plot_embed_clonotype}
\title{plot_embed_clonotype}
\usage{
plot_embed_clonotype(
  input,
  title = "",
  clonotype_id,
  clonotype_by,
  color_by,
  facet_by = NULL,
  ncol = NULL,
  label = FALSE,
  size = 1.5,
  alpha = 0.3,
  colors = NULL,
  theme = "classic",
  legend_dot_size = 1.5,
  xcol = "x",
  ycol = "y",
  reduction = "umap",
  text_sizes = c(20, 10, 5, 10),
  shuffle = F
)
}
\arguments{
\item{input}{ExpressionSet or Seurat Object}

\item{title}{Title of the ggplot2 plot}

\item{clonotype_id}{Clonotype ID among values listed by "clonotype_by" parameter}

\item{clonotype_by}{the label of the column in metadata that defines clonotypes}

\item{color_by}{What to color points by, either "UMI_sum", or pData categorial variable, ignored if gene is provided}

\item{facet_by}{What to break the plots by}

\item{ncol}{How many columns if faceting}

\item{label}{if true, labels of "color_by" will be visualized on embedding}

\item{size}{The size of the points}

\item{alpha}{The transparency of the points}

\item{colors}{What colors to utilize for categorial data. Be sure it is of the proper length!}

\item{theme}{display theme}

\item{legend_dot_size}{Size of dot in legend}

\item{xcol}{pData column to use for x axis}

\item{ycol}{pData column to use for y axis}

\item{reduction}{the reduction key for Seurat objects}

\item{text_sizes}{text sizes for multiple components of the ggplot figure}

\item{shuffle}{change order of labels}
}
\description{
Visual and highlight clonotypes given an embedding
}
\examples{
# library
library(Seurat)

# analyzed data with tcr
data("seu_analyzed")

# get clonotype highlighted on embedding
plot_embed_clonotype(seu_analyzed, clonotype_by = "cdr3s_aa", clonotype_id = "TRB:CASNPSTGGGGEQYF;TRA:CAVMDSNYQLIW",
                     color_by = "seurat_clusters")
plot_embed_clonotype(seu_analyzed, clonotype_by = "cdr3s_aa", clonotype_id = "TRB:CASNPSTGGGGEQYF;TRA:CAVMDSNYQLIW",
                     color_by = "seurat_clusters", reduction = "tsne", label = T)

}
