% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_heatmap_clonotypes}
\alias{plot_heatmap_clonotypes}
\title{plot_heatmap_clonotypes}
\usage{
plot_heatmap_clonotypes(
  input,
  clonotype_by,
  patient_by = NULL,
  sample_by = NULL,
  group_by = NULL,
  ntop = 20,
  Count_limit = NULL
)
}
\arguments{
\item{input}{Seurat Object or ExpressionSet}

\item{clonotype_by}{meta.data or pData column name for clonotype ID's}

\item{patient_by}{meta.data or pData column name for patient}

\item{sample_by}{meta.data or pData column name for sample}

\item{group_by}{What to color points by, either "UMI_sum", or meta.data or pData column name, ignored if gene is provided}

\item{ntop}{the number of top clonotypes}

\item{Count_limit}{count limit for plot legend}
}
\description{
Visual the clonotype expansion in an heatmap
}
\examples{
plot_heatmap_clonotypes(pfizer, clonotype_by = "CTaa", patient_by = "Patient", sample_by = "Sample") # ClonotypeIDs not listed, as not directly comparable across different patients
plot_heatmap_clonotypes(pfizer, clonotype_by = "CTaa", group_by = "Sample_Type")
plot_heatmap_clonotypes(pfizer[,which(pData(pfizer)$Patient=="VB234")], clonotype_by = "CTaa", group_by="Sample")

}
