% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seu_analyzed}
\alias{seu_analyzed}
\title{Human T cells from a Healthy Donor, 1k cells - multi (v2)}
\format{
A Seurat object of 1137 cells and 36601 genes/features
}
\usage{
seu_analyzed
}
\description{
A Seurat single cell object of Human T cells from a Healthy Donor, 1k cells - multi (v2)
}
\keyword{datasets}
