% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{VDJ_visualize}
\alias{VDJ_visualize}
\title{VDJ_visualize}
\usage{
VDJ_visualize(
  df,
  gene = "v_gene",
  chain = "TRA",
  plot = "bar",
  y.axis = "sample",
  order = "gene",
  scale = TRUE,
  split.by = NULL
)
}
\arguments{
\item{gene}{either v_gene, d_gene or j_gene}

\item{chain}{either TRA or TRB}

\item{plot}{only bar}

\item{split.by}{Which meta.data to split the data by.}
}
\value{

}
\description{
Visualize V, D or J gene frequencies across Patients or Samples in a bar graph.
}
