% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{get_topclonotypes}
\alias{get_topclonotypes}
\title{get_topclonotypes}
\usage{
get_topclonotypes(
  input,
  clonotype_by,
  group_by,
  ntop = 20,
  split_by = NULL,
  chain_type = NULL
)
}
\arguments{
\item{input}{ExpressionSet or Seurat Object}

\item{clonotype_by}{the label of the column in metadata that defines clonotypes}

\item{group_by}{what to group clonotypes frequencies}

\item{ntop}{the number of top clonotypes}

\item{split_by}{split tables by}

\item{chain_type}{subset table to chain type}
}
\value{

}
\description{
Get top clonotypes categorized by a column in metadata
}
