% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_heatmap_clonotypes}
\alias{plot_heatmap_clonotypes}
\title{plot_heatmap_clonotypes}
\usage{
plot_heatmap_clonotypes(
  input,
  clonotype_by,
  group_by = NULL,
  sample_by = NULL,
  patient_by = NULL,
  ntop = 20,
  Count_limit = NULL
)
}
\arguments{
\item{input}{ExpressionSet or Seurat Object}

\item{clonotype_by}{the label of the column in metadata that defines clonotypes}

\item{group_by}{What to color points by, either "UMI_sum", or pData categorial variable, ignored if gene is provided}

\item{sample_by}{name of the sample metadata variable}

\item{patient_by}{name of the patient metadata variable}

\item{ntop}{the number of top clonotypes}

\item{Count_limit}{count limit for plot legend}
}
\description{
Visual the clonotype expansion in an heatmap
}
