% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_heatmap_clonotypes}
\alias{plot_heatmap_clonotypes}
\title{plot_heatmap_clonotypes}
\usage{
plot_heatmap_clonotypes(
  input,
  clonotype_by,
  color_by,
  ntop = 20,
  Count_limit = NULL
)
}
\arguments{
\item{input}{ExpressionSet or Seurat Object}

\item{clonotype_by}{the label of the column in metadata that defines clonotypes}

\item{color_by}{What to color points by, either "UMI_sum", or pData categorial variable, ignored if gene is provided}

\item{ntop}{the number of top clonotypes}

\item{Count_limit}{count limit for plot legend}
}
\description{
Visual the clonotype expansion in an heatmap
}
