% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_heatmap_clonotypes}
\alias{plot_heatmap_clonotypes}
\title{plot_heatmap_clonotypes}
\usage{
plot_heatmap_clonotypes(
  input,
  clonotype_by,
  patient_by = NULL,
  sample_by = NULL,
  group_by = NULL,
  sort_by = NULL,
  sort_id = NULL,
  ntop = 20,
  Count_limit = NULL
)
}
\arguments{
\item{input}{Seurat Object or ExpressionSet}

\item{clonotype_by}{meta.data or pData column name for clonotype ID's}

\item{patient_by}{meta.data or pData column name for patient}

\item{sample_by}{meta.data or pData column name for sample}

\item{group_by}{What to color points by, either "UMI_sum", or meta.data or pData column name, ignored if gene is provided}

\item{sort_by}{The metadata column for sorting}

\item{sort_id}{The value in the metadata column for sorting}

\item{ntop}{the number of top clonotypes to return}

\item{Count_limit}{count limit for plot legend}
}
\description{
Visualize clonal expansion by heatmap. Default shows the clonal expansion of the top 20 unique clonotypes across all of a grouping variable,
or the number of clones of the top 20 clonotypes per each facet by multiple grouping variables
Bottom row annotation shows the total number of clones in each grouping variable
}
\examples{
plot_heatmap_clonotypes(pfizer, clonotype_by = "CTaa", patient_by = "Patient", sample_by = "Sample") # ClonotypeIDs not listed, as not directly comparable across different patients
plot_heatmap_clonotypes(pfizer, clonotype_by = "CTaa", group_by = "Sample_Type")
plot_heatmap_clonotypes(pfizer[,which(pData(pfizer)$Patient=="VB234")], clonotype_by = "CTaa", group_by="Sample")

}
